# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Image Recognition Microservices App
  description: The app allows users to run inference on images for image detection and provides additional functionalities like video streaming, user management, and API usage tracking.
  version: v1
servers:
  - url: https://isa-singh.azurewebsites.net/
paths:
  /login:
    get:
      summary: Display the login page
      description: Returns the login page HTML.
      responses:
        '200':
          description: Successfully retrieved the login page.
    post:
      summary: Authenticate user
      description: Authenticates the user with email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully authenticated.
        '401':
          description: Unauthorized.
  /register:
    post:
      summary: Register a new user
      description: Allows a user to register with email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully registered.
        '401':
          description: Unauthorized.
    get:
      summary: Display the registration page
      description: Returns the registration page HTML.
      responses:
        '200':
          description: Successfully retrieved the registration page.
  /forgot-password:
    post:
      summary: Forgot password
      description: Allows a user to request a password reset using their email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: Password reset email sent.
        '401':
          description: Unauthorized.
...
  /detect:
    get:
      summary: Get image with AI inference
      description: Returns a JPEG image with inference lines based on AI detection.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully returned image with inference.
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized.
...
  /video-feed:
    get:
      summary: Live video feed
      description: Returns a live video feed.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully returned live video feed.
        '401':
          description: Unauthorized.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT